+++ lib/Author.php
@@ -1,6 +1,6 @@
 <?php
 declare(strict_types=1);
-namespace akismet;
+namespace Akismet;
 
 /**
  * Represents the author of a comment.
@@ -61,16 +61,15 @@ class Author implements \JsonSerializable {
    * @param mixed $map A JSON map representing an author.
    * @return Author The instance corresponding to the specified JSON map, or `null` if a parsing error occurred.
    */
-  public static function fromJSON($map) {
+  public static function fromJson($map) {
     if (is_array($map)) $map = (object) $map;
     else if (!is_object($map)) return null;
 
-    /** @var Author $author */
     $author = new static(isset($map->user_ip) && is_string($map->user_ip) ? $map->user_ip : '');
     return $author->setEmail(isset($map->comment_author_email) && is_string($map->comment_author_email) ? $map->comment_author_email : '')
       ->setName(isset($map->comment_author) && is_string($map->comment_author) ? $map->comment_author : '')
       ->setRole(isset($map->user_role) && is_string($map->user_role) ? $map->user_role : '')
-      ->setURL(isset($map->comment_author_url) && is_string($map->comment_author_url) ? $map->comment_author_url : '')
+      ->setUrl(isset($map->comment_author_url) && is_string($map->comment_author_url) ? $map->comment_author_url : '')
       ->setUserAgent(isset($map->user_agent) && is_string($map->user_agent) ? $map->user_agent : '');
   }
 
@@ -110,7 +109,7 @@ class Author implements \JsonSerializable {
    * Gets the URL of the author's website.
    * @return string The URL of the author's website.
    */
-  public function getURL(): string {
+  public function getUrl(): string {
     return $this->url;
   }
 
@@ -130,7 +129,7 @@ class Author implements \JsonSerializable {
     $map = new \stdClass;
     if (mb_strlen($name = $this->getName())) $map->comment_author = $name;
     if (mb_strlen($email = $this->getEmail())) $map->comment_author_email = $email;
-    if (mb_strlen($url = $this->getURL())) $map->comment_author_url = $url;
+    if (mb_strlen($url = $this->getUrl())) $map->comment_author_url = $url;
     if (mb_strlen($userAgent = $this->getUserAgent())) $map->user_agent = $userAgent;
     if (mb_strlen($ipAddress = $this->getIPAddress())) $map->user_ip = $ipAddress;
     if (mb_strlen($role = $this->getRole())) $map->user_role = $role;
@@ -182,7 +181,7 @@ class Author implements \JsonSerializable {
    * @param string $value The new website URL.
    * @return Author This instance.
    */
-  public function setURL(string $value): self {
+  public function setUrl(string $value): self {
     $this->url = $value;
     return $this;
   }